+++ library/Haste/Dca/Member.php
@@ -12,13 +12,15 @@
 namespace HeimrichHannot\Haste\Dca;
 
 
+use HeimrichHannot\Haste\Database\QueryHelper;
 use HeimrichHannot\Haste\Util\Arrays;
 
 class Member extends \Backend
 {
-    protected static $arrMemberOptionsCache        = [];
-    protected static $arrMemberOptionsIdsCache     = [];
-    protected static $arrMemberOptionsEmailIdCache = [];
+    protected static $arrMemberOptionsCache                = [];
+    protected static $arrMemberOptionsIdsCache             = [];
+    protected static $arrMemberOptionsEmailIdCache         = [];
+    protected static $arrMemberOptionsEmailIdCacheByGroups = [];
 
     public static function getMembersAsOptions(\DataContainer $objDc = null, $blnIncludeId = false)
     {
@@ -73,7 +75,7 @@ class Member extends \Backend
         return static::getMembersAsOptions($objDc, true);
     }
 
-    public static function getMembersAsOptionsIncludingEmailAndId(\DataContainer $objDc = null, $blnIncludeId = false)
+    public static function getMembersAsOptionsIncludingEmailAndId(\DataContainer $objDc = null)
     {
         if (!empty(static::$arrMemberOptionsEmailIdCache))
         {
@@ -108,4 +110,42 @@ class Member extends \Backend
         return $arrOptions;
     }
 
+    public static function getMembersAsOptionsIncludingEmailAndIdByGroups(\DataContainer $objDc = null, array $arrGroups)
+    {
+        $strGroupKey = implode('_', $arrGroups);
+
+        if (!empty(static::$arrMemberOptionsEmailIdCacheByGroups[$strGroupKey]))
+        {
+            return static::$arrMemberOptionsEmailIdCacheByGroups[$strGroupKey];
+        }
+
+        $objDatabase = \Database::getInstance();
+        $objMembers  = $objDatabase->execute(
+            'SELECT id, firstname, lastname, email, groups FROM tl_member WHERE ' . QueryHelper::createWhereForSerializedBlob('groups', $arrGroups)
+        );
+        $arrOptions  = [];
+
+        if ($objMembers->numRows > 0)
+        {
+            while ($objMembers->next())
+            {
+                $arrAttributes = [];
+
+                if ($objMembers->email)
+                {
+                    $arrAttributes[] = $objMembers->email;
+                }
+
+                $arrAttributes[] = 'ID ' . $objMembers->id;
+
+                $arrOptions[$objMembers->id] = $objMembers->lastname . ', ' . $objMembers->firstname . ' (' . implode(', ', $arrAttributes) . ')';
+            }
+        }
+
+        asort($arrOptions);
+
+        $arrMemberOptionsEmailIdCacheByGroups[$strGroupKey] = $arrOptions;
+
+        return $arrOptions;
+    }
 }